pal:do not set the initial volume 0 for LL in MSPP
MSPP volume is set to 0 during intialization and we are not updating the
actual volume for touch tones during the usecase start. Hence the touch
tones are muted on the MSPP.This was same even with earlier SPs however
due to latest changes in MSPP_v5 lib,in which volume is more sensitive
and uncovered this bug.

Therefore we are not seting the initial volume 0 for LL stream from
HLOS.

Signed-off-by: Chiluka Rohith <rchiluka@qti.qualcomm.com>