PAL: Refine sleep monitor voting mechanism
Previously, voting was done based on static decision, if one stream is
classified as low power it will vote for low power, irrespective of
usecase state.

Now, updated the logic, to use current usecase state to vote, if usecase
is running in low power then it will vote for low power, otherwise it
will vote for high power.

Signed-off-by: Chiluka Rohith <rchiluka@qti.qualcomm.com>